Spreadsheet with tab for Dynamic P&L statement
I was tasked with creating a dashboard for a Profit and Loss Statement that calculates the Gross Profit, Operating Profit, or EBIT for a company selected from a drop-down list. The details that were needed in the spreadsheet included: all companies from the raw data, drop-down list needed to pull historical fundamentals data to create the P&L Statement, and the P&L statement needed to include the Gross Profit, Operating Profit, or EBIT values for all the years there is historical data available for that company in the dataset.